Move bootstrap to its own module
